devprofile-resume-automator/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
supabase==2.3.0
python-multipart==0.0.6
jinja2==3.1.2
weasyprint==60.2
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.26.0